Change int to long

+++ src/ResourceManager/OperationalInsights/Commands.OperationalInsights/Search/GetAzureOperationalInsightsSearchResultsCommand.cs
@@ -35,7 +35,7 @@ namespace Microsoft.Azure.Commands.OperationalInsights
         [Parameter(Position = 2, Mandatory = false, ValueFromPipelineByPropertyName = true,
             HelpMessage = "The top search parameter.")]
         [ValidateNotNullOrEmpty]
-        public int Top { get; set; }
+        public long Top { get; set; }
 
         [Parameter(Position = 3, Mandatory = false, ValueFromPipelineByPropertyName = true,
             HelpMessage = "The pre highlight search parameter.")]
